backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
weasyprint==60.2
pydyf==0.10.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
Pillow==10.4.0

